User is now truely logged in, can only see their saved retailers

+++ src/main/java/com/seanmiller/repository/UserRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import com.seanmiller.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String username);
-    
+        
     boolean existsByUsername(String username);
     
     boolean existsByUsernameAndPassword(String username, String password);
